+++ contracts/src/events/purchased.cairo
@@ -2,7 +2,7 @@ pub use crate::events::index::Purchased;
 
 #[generate_trait]
 pub impl PurchasedImpl of PurchasedTrait {
-    fn new(player_id: felt252, starterpack_id: u32, quantity: u32, multiplier: u8) -> Purchased {
+    fn new(player_id: felt252, starterpack_id: u32, quantity: u32, multiplier: u128) -> Purchased {
         Purchased {
             player_id: player_id,
             starterpack_id: starterpack_id,
